+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/psi/stub/index/VariableNameIndex.java
@@ -39,6 +39,6 @@ public class VariableNameIndex extends StringStubIndexExtension<RobotVariable> {
 
     @Override
     public int getVersion() {
-        return super.getVersion() + 4;
+        return super.getVersion() + 5;
     }
 }