+++ src/main/kotlin/dev/xeonkryptos/xeonrobotframeworkplugin/formatter/RobotCodeStyleSettings.kt
@@ -3,9 +3,12 @@ package dev.xeonkryptos.xeonrobotframeworkplugin.formatter
 import com.intellij.psi.codeStyle.CodeStyleSettings
 import com.intellij.psi.codeStyle.CustomCodeStyleSettings
 
+@Suppress("PropertyName")
 class RobotCodeStyleSettings(container: CodeStyleSettings) : CustomCodeStyleSettings("RobotCodeStyleSettings", container) {
-    @JvmField var ALIGN_SETTINGS = true
-    @JvmField var ALIGN_VARIABLES = true
-    @JvmField var ALIGN_TEST_CASES = true
-    @JvmField var ALIGN_KEYWORDS = true
+    companion object {
+        const val SUPER_SPACE_SIZE = 2
+    }
+
+    @JvmField
+    var SPACE_AROUND_VARIABLE_BODY: Boolean = false
 }